% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNewID.R
\docType{methods}
\name{createNewID}
\alias{createNewID}
\alias{createNewID,sdcMicroObj,character,characterOrNULL-method}
\title{Creates new randomized IDs}
\usage{
createNewID(obj, newID, withinVar)

\S4method{createNewID}{sdcMicroObj,character,characterOrNULL}(obj, newID,
  withinVar = NULL)
}
\arguments{
\item{obj}{an object of class \code{sdcMicroObj}}

\item{newID}{a character specifiying the desired variable name of the new ID}

\item{withinVar}{if not \code{NULL} a character vector specifying a variable (e.g an existing household ID) which
will be used when calculating the new IDs. If specified, the same IDs will be assigned to the same values of the given variable.}
}
\value{
an object of class \code{sdcMicroObj} with updated slot \code{origData}
}
\description{
This is useful if the record IDs consist, for example, of a geo identifier and the household line number.
This method can be used to create new, random IDs that cannot be reconstructed.
}

