% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distributeRandom_cpp}
\alias{distributeRandom_cpp}
\title{Distribute}
\usage{
distributeRandom_cpp(inputRatio, totalDraws, seed)
}
\arguments{
\item{inputRatio}{vector containing ratios which are used to distribute number units in `totalDraws`.}

\item{totalDraws}{number of units to distribute}

\item{seed}{integer setting the sampling seed}
}
\description{
Distribute `totalDraws` using ratio/probability vector `inputRatio` and randomly round each entry up or down such that the distribution results in an integer vector.
Returns an integer vector containing the number of units in `totalDraws` distributetd according to proportions in `inputRatio`.
\cr
\strong{NOTE:} This is an internal function used for testing the C++-function \code{distributeRandom} which is used inside the C++-function \code{recordSwap()}.
}
