% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{get.sdcMicroObj}
\alias{get.sdcMicroObj}
\title{get.sdcMicroObj}
\usage{
get.sdcMicroObj(object, type)
}
\arguments{
\item{object}{a \code{\link{sdcMicroObj-class}}-object}

\item{type}{a character vector of length 1 defining what to calculate|return|modify. Allowed types are are
all slotNames of \code{obj}.}
}
\value{
a slot of a \code{\link{sdcMicroObj-class}}-object depending on argument \code{type}
}
\description{
extract information from \code{\link{sdcMicroObj-class}}-objects depending on argument \code{type}
}
\examples{
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
sl <- slotNames(sdc)
res <- sapply(sl, function(x) get.sdcMicroObj(sdc, type=x))
str(res)
}
