% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIfunctions.R
\docType{methods}
\name{groupAndRename}
\alias{groupAndRename}
\alias{groupAndRename,data.frame-method}
\alias{groupAndRename,data.frame-method,}
\alias{groupAndRename,factor-method}
\alias{groupAndRename,factor-method,groupAndRename,sdcMicroObj-method}
\alias{groupAndRename,sdcMicroObj-method}
\title{Join levels of a variables in an object of class
\code{\link{sdcMicroObj-class}} or \code{factor} or \code{data.frame}}
\usage{
groupAndRename(obj, var, before, after)

\S4method{groupAndRename}{factor}(obj, var, before, after)

\S4method{groupAndRename}{data.frame}(obj, var, before, after)

\S4method{groupAndRename}{sdcMicroObj}(obj, var, before, after)
}
\arguments{
\item{obj}{object of class \code{\link{sdcMicroObj-class}}}

\item{var}{name of the keyVariable to change}

\item{before}{vector of levels before recoding}

\item{after}{vector of levels after recoding}
}
\value{
the modified \code{\link{sdcMicroObj-class}}
}
\description{
If the input is an object of class \code{\link{sdcMicroObj-class}}, the
specified factor-variable is recoded into a factor with less levels and
risk-measures are automatically recomputed.
}
\details{
If the input is of class \code{data.frame}, the result is a \code{data.frame} with
a modified column specified by \code{var}.

If the input is of class \code{factor}, the result is a \code{factor} with different
levels.
}
\section{Methods}{
 \describe{
\item{list("signature(obj = \"sdcMicroObj\")")}{
This method transform a factor variable with some levels into a new factor
variable with less levels. The user must make sure that all levels of the
original variable are listed in argument 'before' and that the number of
elements in argument 'after' (the new levels) have the same length. This
means that there should be a one to one mapping from any level of the
original factor to a level in the recoded variable. } }
}
\examples{
## for objects of class sdcMicro:
data(testdata2)
testdata2$urbrur <- as.factor(testdata2$urbrur)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
sdc <- groupAndRename(sdc, var="urbrur", before=c("1","2"), after=c("1"))
}
\author{
Bernhard Meindl
}
\keyword{methods}

