% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIfunctions.R
\name{importProblem}
\alias{importProblem}
\title{importProblem}
\usage{
importProblem(path)
}
\arguments{
\item{path}{a file path}
}
\value{
an object of class \code{sdcMicro_GUI_export} or an object of class 'simple.error'
}
\description{
reads an sdcProblem with code that has been exported within \code{\link{sdcApp}}.
}
\author{
Bernhard Meindl
}

