% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{kAnon_violations}
\alias{kAnon_violations}
\alias{kAnon_violations,sdcMicroObj,logical,numeric-method}
\title{\code{kAnon_violations}}
\usage{
kAnon_violations(object, weighted, k)

\S4method{kAnon_violations}{sdcMicroObj,logical,numeric}(object, weighted, k)
}
\arguments{
\item{object}{a \code{\link{sdcMicroObj-class}} object}

\item{weighted}{\code{TRUE} or \code{FALSE} defining if sampling weights should be taken into account}

\item{k}{a positive number defining parameter k}
}
\value{
the number of records that are violating k-anonymity based on 
unweighted sample data only (in case parameter \code{weighted} is \code{FALSE}) or computing
the number of observations that are estimated to violate k-anonymity in the population in case 
parameter \code{weighted} equals \code{TRUE}.
}
\description{
returns the number of observations violating k-anonymity.
}
