% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTopCoding.R
\name{mvTopCoding}
\alias{mvTopCoding}
\title{Detection and winsorization of multivariate outliers}
\usage{
mvTopCoding(x, maha=NULL,center=NULL,cov=NULL, alpha=0.025)
}
\arguments{
\item{x}{data frame or matrix or table}

\item{maha}{squared mahalanobis distance of each observation}

\item{center}{center of data, needed for calcualtion of mahalanobis distance (if not provide)}

\item{alpha}{significance level, determining the ellipsoide to which outliers should be placed upon}

\item{sigma}{covariance matrix of data, needed for calcualtion of mahalanobis distance (if not provide)}
}
\value{
the imputed winsorized data
}
\description{
Imputation and detection of outliers
}
\details{
Winsorizes the potential outliers on the ellipsoid defined by MD2 in direction to the center of the data
}
\examples{
set.seed(123)
x <- MASS::mvrnorm(20, mu = c(5,5), Sigma = matrix(c(1,0.9,0.9,1), ncol = 2))
x[1,1] <- 3
x[1,2] <- 6
plot(x)
ximp <- mvTopCoding(x)
}
\author{
Johannes Gussenbauer, Matthias Templ
}
