% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.r
\docType{methods}
\name{nextSdcObj}
\alias{nextSdcObj}
\title{nextSdcObj}
\usage{
nextSdcObj(obj)
}
\arguments{
\item{obj}{a \code{\link{sdcMicroObj-class}} object}
}
\value{
a modified \code{\link{sdcMicroObj-class}} object
}
\description{
internal function used to provide the undo-functionality.
}
