% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orderData_cpp}
\alias{orderData_cpp}
\title{Reorder data}
\usage{
orderData_cpp(data, orderIndex)
}
\arguments{
\item{data}{micro data set containing only numeric values.}

\item{orderIndex}{column index in \code{data} refering to the column by which data should be ordered.}
}
\value{
ordered data set.
}
\description{
Reorders the data according to a column in the data set.
\cr
\strong{NOTE:} This is an internal function used for testing the C++-function \code{orderData} which is used inside the C++-function \code{recordSwap()} to speed up performance.
}
