% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivRisk.R
\name{print.indivRisk}
\alias{print.indivRisk}
\title{Print method for objects from class indivRisk}
\usage{
\method{print}{indivRisk}(x, ...)
}
\arguments{
\item{x}{object from class indivRisk}

\item{\dots}{Additional arguments passed through.}
}
\value{
few information about the method and the final correction factor for
objects of class \sQuote{indivRisk}.
}
\description{
Print method for objects from class indivRisk
}
\examples{

## example from Capobianchi, Polettini and Lucarelli:
data(francdat)
f1 <- freqCalc(francdat, keyVars=c(2,4,5,6),w=8)
data.frame(fk=f1$fk, Fk=f1$Fk)
## individual risk calculation:
indivRisk(f1)

}
\seealso{
\code{\link{indivRisk}}
}
\author{
Matthias Templ
}
\keyword{print}
