% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randSample_cpp}
\alias{randSample_cpp}
\title{Random Sampling}
\usage{
randSample_cpp(ID, N, prob, IDused, seed)
}
\arguments{
\item{ID}{vector containing record IDs from which to sample}

\item{N}{integer defining the number of records to be sampled}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}

\item{IDused}{vector containing IDs which must not be sampled}

\item{seed}{integer setting the sampling seed}
}
\description{
Randomly select records given a probability weight vector \code{prob}.
\cr
\strong{NOTE:} This is an internal function used for testing the C++-function \code{randSample} which is used inside the C++-function \code{recordSwap()}.
}
