% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIfunctions.R
\docType{methods}
\name{removeDirectID}
\alias{removeDirectID}
\title{Remove certain variables from the data set inside a sdc object.}
\usage{
removeDirectID(obj, var)
}
\arguments{
\item{obj}{object of class \code{\link{sdcMicroObj-class}}}

\item{var}{name of the variable(s) to be remove}
}
\value{
the modified \code{\link{sdcMicroObj-class}}
}
\description{
Delete variables without changing anything else in the sdcObject (writing
NAs).
}
\examples{
## for objects of class sdcMicro:
data(testdata2)
sdc <- createSdcObj(testdata, keyVars=c('urbrur','roof'),
 numVars=c('expend','income','savings'), w='sampling_weight')
sdc <- removeDirectID(sdc, var="age")
}
\author{
Alexander Kowarik
}
\keyword{methods}

