% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcApp.R
\name{sdcApp}
\alias{sdcApp}
\title{sdcApp}
\usage{
sdcApp(
  maxRequestSize = 50,
  debug = FALSE,
  theme = "IHSN",
  ...,
  shiny.server = FALSE
)
}
\arguments{
\item{maxRequestSize}{(numeric) number defining the maximum allowed filesize (in megabytes)
for uploaded files, defaults to 50MB}

\item{debug}{logical if \code{TRUE}, set shiny-debugging options}

\item{theme}{select stylesheet for the interface. Supported choices are
\itemize{
\item 'yeti'
\item 'flatly'
\item 'journal'
\item 'IHSN'
}}

\item{...}{arguments (e.g \code{host}) that are passed through \code{\link[shiny]{runApp}} when
starting the shiny application}

\item{shiny.server}{Setting this parameter to \code{TRUE} will return the app in the form of an
object rather than invoking it. This is useful for deploying \code{sdcApp} via \code{shiny-server}.}
}
\value{
starts the interactive graphical user interface which may be used to perform the
anonymisation process.
}
\description{
starts the graphical user interface developed with \emph{shiny}.
}
\examples{
if(interactive()) {
  sdcApp(theme = "flatly")
}
}
