% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcGUI.R
\name{sdcGUI}
\alias{sdcGUI}
\title{sdcGUI}
\usage{
sdcGUI()
}
\value{
starts the interactive graphical user interface which may be used to perform the
anonymisation process.
}
\description{
sdcGUI
}
\examples{
\dontrun{
sdcGUI()
}
}

