% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sdcMicroObj.R
\name{show,sdcMicroObj-method}
\alias{show,sdcMicroObj-method}
\title{Show}
\usage{
\S4method{show}{sdcMicroObj}(object)
}
\arguments{
\item{object}{an sdcmicro obj}
}
\value{
a sdcMicro object
}
\description{
show a sdcMicro object
}
\author{
Bernhard Meindl
}
