\name{testdata}
\alias{testdata}
\docType{data}
\title{
Dataset for testing purpose
}
\description{
A concise (1-5 lines) description of the dataset.
}
\usage{data(testdata)}
\format{
  A data frame with 4580 observations on the following 14 variables.
  \describe{
    \item{\code{urbrur}}{a numeric vector}
    \item{\code{roof}}{a numeric vector}
    \item{\code{walls}}{a numeric vector}
    \item{\code{water}}{a numeric vector}
    \item{\code{electcon}}{a numeric vector}
    \item{\code{relat}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{hhcivil}}{a numeric vector}
    \item{\code{expend}}{a numeric vector}
    \item{\code{income}}{a numeric vector}
    \item{\code{savings}}{a numeric vector}
    \item{\code{ori_hid}}{a numeric vector}
    \item{\code{sampling_weight}}{a numeric vector}
  }
}
\details{
If necessary, more details than the __description__ above
}
\source{
reference to a publication or URL from which the data were obtained
}
\references{
possibly secondary sources and usages
}
\examples{
data(testdata)
## maybe str(testdata) ; plot(testdata) ...
}
\keyword{datasets}