% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIfunctions.R
\docType{methods}
\name{varToFactor}
\alias{varToFactor}
\alias{varToFactor,sdcMicroObj-method}
\alias{varToFactor-methods}
\alias{varToNumeric}
\alias{varToNumeric,sdcMicroObj-method}
\alias{varToNumeric-methods}
\title{Change the a keyVariable of an object of class \code{\link{sdcMicroObj-class}} from Numeric to
Factor or from Factor to Numeric}
\usage{
varToFactor(obj, var)
}
\arguments{
\item{obj}{object of class \code{\link{sdcMicroObj-class}}}

\item{var}{name of the keyVariable to change}
}
\value{
the modified \code{\link{sdcMicroObj-class}}
}
\description{
Change the scale of a variable
}
\section{Methods}{
 \describe{
\item{list("signature(obj = \"sdcMicroObj\")")}{}}
}
\examples{

## for objects of class sdcMicro:
data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
sdc <- varToFactor(sdc, var="urbrur")

}
\keyword{methods}

