\name{casc1}
\alias{casc1}
\docType{data}
\title{ Small Artificial Data set }
\description{
  Small Toy Example Data set which was used by Sanz-Mateo et.al.
}
\usage{data(casc1)}
\format{
  The format is:
 int [1:13, 1:7] 10 12 17 21 9 12 12 14 13 15 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:13] "1" "2" "3" "4" ...
  ..$ : chr [1:7] "1" "2" "3" "4" ...
}
\examples{
data(casc1)
casc1
}
\keyword{datasets}
