\name{CASCrefmicrodata}
\alias{CASCrefmicrodata}
\docType{data}
\title{ Census data set }
\description{
The test data set was obtained on July 27, 2000 using the Data Extraction System of the U. S. Bureau of the Census.
}
\usage{data(CASCrefmicrodata)}
\format{
  A data frame with 1080 observations on the following 13 variables.
  \describe{
    \item{\code{AFNLWGT}}{a numeric vector}
    \item{\code{AGI}}{a numeric vector}
    \item{\code{EMCONTRB}}{a numeric vector}
    \item{\code{FEDTAX}}{a numeric vector}
    \item{\code{PTOTVAL}}{a numeric vector}
    \item{\code{STATETAX}}{a numeric vector}
    \item{\code{TAXINC}}{a numeric vector}
    \item{\code{POTHVAL}}{a numeric vector}
    \item{\code{INTVAL}}{a numeric vector}
    \item{\code{PEARNVAL}}{a numeric vector}
    \item{\code{FICA}}{a numeric vector}
    \item{\code{WSALVAL}}{a numeric vector}
    \item{\code{ERNVAL}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://neon.vb.cbs.nl/casc/testsets.html}
}
\references{
Brand, R. and
Domingo-Ferrer, J. and Mateo-Sanz, J.M., Reference data sets to test and compare {SDC} methods for protection of numerical microdata.
}
\examples{
data(CASCrefmicrodata)
}
\keyword{datasets}
