% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{CASCrefmicrodata}
\alias{CASCrefmicrodata}
\title{Census data set}
\format{A data frame sampled from year 1995 with 1080 observations on the
following 13 variables.  \describe{
\item{AFNLWGT}{Final weight (2 implied decimal places)}
\item{AGI}{Adjusted gross income}
\item{EMCONTRB}{Employer contribution for hlth insurance}
\item{FEDTAX}{Federal income tax liability}
\item{PTOTVAL}{Total person income}
\item{STATETAX}{State income tax liability}
\item{TAXINC}{Taxable income amount}
\item{POTHVAL}{Total other persons income}
\item{INTVAL}{Amt of interest income}
\item{PEARNVAL}{Total person earnings}
\item{FICA}{Soc. sec. retirement payroll deduction}
\item{WSALVAL}{Amount: Total Wage and salary}
\item{ERNVAL}{Business or Farm net earnings}}}
\source{
Public use file from the CASC project.  More information on this
test data can be found in the paper listed below.
}
\description{
This test data set was obtained on July 27, 2000 using the public use Data
Extraction System of the U.S. Bureau of the Census.
}
\examples{

data(CASCrefmicrodata)
str(CASCrefmicrodata)

}
\references{
Brand, R. and Domingo-Ferrer, J. and Mateo-Sanz, J.M., Reference
data sets to test and compare SDC methods for protection of numerical
microdata.  Unpublished.
\url{http://neon.vb.cbs.nl/casc/CASCrefmicrodata.pdf}
}
\keyword{datasets}

