% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRisk.R
\name{dRisk}
\alias{dRisk}
\title{overal disclosure risk}
\usage{
dRisk(obj, ...)
}
\arguments{
\item{obj}{a \code{data.frame} or object of class \code{\link{sdcMicroObj-class}}}

\item{...}{possible arguments are:
\itemize{
\item {\code{xm}: }{perturbed data}
\item {\code{k}: }{percentage of the standard deviation}}}
}
\value{
The disclosure risk or/and the modified \code{\link{sdcMicroObj-class}}
}
\description{
Distance-based disclosure risk estimation via standard deviation-based
intervals around observations.
}
\details{
An interval (based on the standard deviation) is built around each value of
the perturbed value.  Then we look if the original values lay in these
intervals or not. With parameter k one can enlarge or down scale the
interval.
}
\examples{
data(free1)
free1 <- as.data.frame(free1)
m1 <- microaggregation(free1[, 31:34], method="onedims", aggr=3)
m2 <- microaggregation(free1[, 31:34], method="pca", aggr=3)
dRisk(obj=free1[, 31:34], xm=m1$mx)
dRisk(obj=free1[, 31:34], xm=m2$mx)
dUtility(obj=free1[, 31:34], xm=m1$mx)
dUtility(obj=free1[, 31:34], xm=m2$mx)

## for objects of class sdcMicro:
data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
## this is already made internally: sdc <- dRisk(sdc)
## and already stored in sdc
}
\author{
Matthias Templ
}
\references{
see method SDID in
\url{http://vneumann.etse.urv.es/webCrises/publications/isijcr/lncs3050Outlier.pdf}
}
\seealso{
\code{\link{dUtility}}
}
\keyword{manip}

