\name{francdat}
\alias{francdat}
\docType{data}
\title{ data from the casc project}
\description{
 Small synthetic data from Capobianchi, Polettini, Lucarelli
}
\usage{data(francdat)}
\format{
  A data frame with 8 observations on the following 8 variables.
  \describe{
    \item{\code{Num1}}{a numeric vector}
    \item{\code{Key1}}{a numeric vector}
    \item{\code{Num2}}{a numeric vector}
    \item{\code{Key2}}{a numeric vector}
    \item{\code{Key3}}{a numeric vector}
    \item{\code{Key4}}{a numeric vector}
    \item{\code{Num3}}{a numeric vector}
    \item{\code{w}}{a numeric vector}
  }
}
\details{
}
\source{
\url{http://neon.vb.cbs.nl/casc/Deliv/12d1.pdf}
}
\references{
}
\examples{
data(francdat)
francdat
}
\keyword{datasets}
