% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{francdat}
\alias{francdat}
\title{data from the casc project}
\format{
A data frame with 8 observations on the following 8 variables.
\describe{
\item{Num1}{a numeric vector}
\item{Key1}{Key variable 1. A numeric vector}
\item{Num2}{a numeric vector}
\item{Key2}{Key variable 2. A numeric vector}
\item{Key3}{Key variable 3. A numeric vector}
\item{Key4}{Key variable 4. A numeric vector}
\item{Num3}{a numeric vector}
\item{w}{The weight vector. A numeric vector}}
}
\source{
\url{https://research.cbs.nl/casc/deliv/12d1.pdf}
}
\description{
Small synthetic data from Capobianchi, Polettini, Lucarelli
}
\details{
This data set is very similar to that one which are used by the authors of
the paper given below. We need this data set only for demonstration effect,
i.e. that the package provides the same results as their software.
}
\examples{

data(francdat)
francdat
}
\keyword{datasets}
