% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{free1}
\alias{free1}
\title{Demo data set from mu-Argus}
\format{The format is: num [1:4000, 1:34] 36 36 36 36 36 36 36 36 36 36 ...
- attr(*, "dimnames")=List of 2 ..$ : NULL ..$ : chr [1:34] "REGION" "SEX"
"AGE" "MARSTAT" ...}
\source{
Public use file from the CASC project.
}
\description{
The public use toy demo data set from the mu-Argus software for SDC.
}
\details{
Please, see at the link given below. Please note, that the correlation
structure of the data is not very realistic, especially concerning the
continuous scaled variables which drawn independently from are a
multivariate uniform distribution.
}
\examples{

data(free1)
head(free1)

}
\keyword{datasets}
