% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIfunctions.R
\name{generateStrata}
\alias{generateStrata}
\title{Generate one strata variable from multiple factors}
\usage{
generateStrata(df, stratavars, name)
}
\arguments{
\item{df}{a data.frame}

\item{stratavars}{character vector with variable name}

\item{name}{name of the newly generated variable}
}
\value{
The original data set with one new column.
}
\description{
For strata defined by multiple variables (e.g. sex,age,country) one combined
variable is generated.
}
\examples{

x <- testdata
x <- generateStrata(x,c("sex","urbrur"),"strataIDvar")
head(x)

}
\author{
Alexander Kowarik
}

