% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mafast.R
\docType{methods}
\name{mafast}
\alias{mafast}
\alias{mafast,ANY-method}
\alias{mafast,data.frame-method}
\alias{mafast,matrix-method}
\alias{mafast,sdcMicroObj-method}
\alias{mafast-methods}
\title{Fast and Simple Microaggregation}
\usage{
mafast(obj, variables = NULL, by = NULL, aggr = 3, measure = mean)
}
\arguments{
\item{obj}{either an object of class sdcMicroObj or a data frame or matrix}

\item{variables}{variables to microaggregate. If obj is of class sdcMicroObj
the numerical key variables are chosen per default.}

\item{by}{grouping variable for microaggregation. If obj is of class
sdcMicroObj the strata variables are chosen per default.}

\item{aggr}{aggregation level (default=3)}

\item{measure}{aggregation statistic, mean, median, trim, onestep (default =
mean)}
}
\value{
If \sQuote{obj} was of class \code{\link{sdcMicroObj-class}} the corresponding
slots are filled, like manipNumVars, risk and utility.  If \sQuote{obj} was
of class \dQuote{data.frame} or \dQuote{matrix} an object of the same class
is returned.
}
\description{
Function to perform a fast and simple (primitive) method of
microaggregation. (for large datasets)
}
\section{Methods}{
 \describe{
\item{list("signature(obj = \"ANY\")")}{}
\item{list("signature(obj = \"data.frame\")")}{}
\item{list("signature(obj = \"matrix\")")}{}
\item{list("signature(obj = \"sdcMicroObj\")")}{}}
}
\examples{
data(Tarragona)
m1 <- mafast(Tarragona, variables=c("GROSS.PROFIT","OPERATING.PROFIT","SALES"),aggr=3)
data(testdata)
m2 <- mafast(testdata,variables=c("expend","income","savings"),aggr=50,by="sex")
summary(m2)

## for objects of class sdcMicro:
data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
sdc <- dRisk(sdc)
sdc@risk$numeric
sdc1 <- mafast(sdc,aggr=4)
sdc1@risk$numeric

sdc2 <- mafast(sdc,aggr=10)
sdc2@risk$numeric
\dontrun{
### Performance tests
x <- testdata
for(i in 1:20){
  x <- rbind(x,testdata)
}
system.time(xx <- mafast(x,variables=c("expend","income","savings"),aggr=50,by="sex"))
}
}
\author{
Alexander Kowarik
}
\seealso{
\code{\link{microaggregation}}
}
\keyword{manip}

