% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{microData}
\alias{microData}
\title{microData}
\format{
The format is: num [1:13, 1:5] 5 7 2 1 7 8 12 3 15 4 ...  - attr(*,
"dimnames")=List of 2 ..$ : chr [1:13] "10000" "11000" "12000" "12100" ...
..$ : chr [1:5] "one" "two" "three" "four" ...
}
\description{
Small aritificial toy data set.
}
\examples{

data(microData)
microData <- as.data.frame(microData)
m1 <- microaggregation(microData, method="mdav")
summary(m1)

}
\keyword{datasets}
