% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSuppression.R
\name{plot.localSuppression}
\alias{plot.localSuppression}
\title{Plots for localSuppression objects}
\usage{
\method{plot}{localSuppression}(x, ...)
}
\arguments{
\item{x}{object of derived from \code{\link[=localSuppression]{localSuppression()}}}

\item{...}{Additional arguments, currently available are:
\itemize{
\item \code{"showDetails"}: logical, if set, a plot of suppressions by
strata is shown (if possible)
}}
}
\value{
a \code{ggplot} plot object
}
\description{
This function creates barplots to display the number of suppressed values
in categorical key variables to achieve \code{k}-anonymity.
}
\examples{
data(francdat)
}
\seealso{
\code{\link[=localSuppression]{localSuppression()}}
}
\author{
Bernhard Meindl, Matthias Templ
}
\keyword{plot}
