% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSuppression.R
\name{print.localSuppression}
\alias{print.localSuppression}
\title{Print method for objects from class localSuppression}
\usage{
\method{print}{localSuppression}(x, ...)
}
\arguments{
\item{x}{object from class localSuppression}

\item{\dots}{Additional arguments passed through.}
}
\value{
Information about the frequency counts for key variables for object
of class \sQuote{localSuppression}.
}
\description{
Print method for objects from class localSuppression
}
\examples{
## example from Capobianchi, Polettini and Lucarelli:
data(francdat)
l1 <- localSuppression(francdat, keyVars=c(2,4,5,6))
l1

}
\author{
Matthias Templ
}
\seealso{
\code{\link{localSuppression}}
}
\keyword{print}

