% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modRisk.R
\name{print.modrisk}
\alias{modrisk}
\alias{print.modrisk}
\title{Print method for objects from class modrisk}
\usage{
\method{print}{modrisk}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{modrisk}}}

\item{\dots}{Additional arguments passed through.}
}
\value{
Output of model-based risk estimation
}
\description{
Print method for objects from class modrisk
}
\author{
Bernhard Meindl
}
\seealso{
\code{\link{modRisk}}
}
\keyword{print}

