% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pram.R
\name{print.pram}
\alias{print.pram}
\title{Print method for objects from class pram}
\usage{
\method{print}{pram}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{pram}}}

\item{\dots}{Additional arguments passed through.}
}
\value{
absolute and relative frequencies of changed observations in each modified variable
}
\description{
Print method for objects from class pram
}
\author{
Bernhard Meindl, Matthias Templ

Matthias Templ and Bernhard Meindl
}
\seealso{
\code{\link{pram}}
}
\keyword{print}

