% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suda2.R
\name{print.suda2}
\alias{print.suda2}
\title{Print method for objects from class suda2}
\usage{
\method{print}{suda2}(x, ...)
}
\arguments{
\item{x}{an object of class suda2}

\item{\dots}{additional arguments passed through.}
}
\value{
Table of dis suda scores.
}
\description{
Print method for objects from class suda2.
}
\examples{
\dontrun{
data(testdata)
data_suda2 <- suda2(testdata,variables=c("urbrur","roof","walls","water","sex"))
data_suda2
}
}
\author{
Matthias Templ
}
\seealso{
\code{\link{suda2}}
}
\keyword{print}

