% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{report}
\alias{report}
\alias{report,sdcMicroObj-method}
\alias{report-methods}
\title{Generate a HTML/LATEX output from an sdcMicroObj}
\usage{
report(obj, outdir = getwd(), filename = "SDC-Report", format = "HTML",
  title = "SDC-Report", internal = FALSE)
}
\arguments{
\item{obj}{an object of class \code{\link{sdcMicroObj-class}} or 'reportObj'}

\item{outdir}{output folder}

\item{filename}{output filename}

\item{format}{HTML, TEXT or LATEX}

\item{title}{Title for the report}

\item{internal}{TRUE/FALSE, if TRUE a detailled internal report is produced,
else a non-disclosive overview}
}
\description{
Summary statistics of the original and the perturbed data set
}
\details{
The application of this function provides you with a html, text or
pdf-report for your sdcMicro object that contains useful summaries about the
anonymization process.
}
\examples{
\dontrun{
data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
report(sdc)
}
}
\author{
Matthias Templ, Bernhard Meindl
}
\keyword{methods}

