% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcApp.R
\name{sdcApp}
\alias{sdcApp}
\title{sdcApp}
\usage{
sdcApp(maxRequestSize = 50, debug = FALSE, theme = "yeti")
}
\arguments{
\item{maxRequestSize}{(numeric) number defining the maximum allowed filesize (in megabytes)
for uploaded files, defaults to 50MB}

\item{debug}{logical if \code{TRUE}, set shiny-debugging options}

\item{theme}{select stylesheet for the interface. Supported choices are
\itemize{
\item 'yeti'
\item 'flatly'
\item 'journal'
}}
}
\value{
starts the interactive graphical user interface which may be used to perform the
anonymisation process.
}
\description{
starts the graphical user interface developed with \emph{shiny}.
}
\examples{
\dontrun{
sdcApp(theme="flatly")
}
}
