% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcGUI.R
\name{sdcGUI}
\alias{sdcGUI}
\title{sdcGUI}
\usage{
sdcGUI(maxRequestSize = 50)
}
\arguments{
\item{maxRequestSize}{(numeric) number defining the maximum allowed filesize (in megabytes)
for uploaded files, defaults to 50MB}
}
\value{
starts the interactive graphical user interface which may be used to perform the
anonymisation process.
}
\description{
starts the graphical user interface developed with \emph{shiny}.
}
\examples{
\dontrun{
sdcGUI()
}
}

