% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{Tarragona}
\alias{Tarragona}
\title{Tarragona data set}
\format{A data frame with 834 observations on the following 13 variables.
\describe{
\item{FIXED.ASSETS}{a numeric vector}
\item{CURRENT.ASSETS}{a numeric vector}
\item{TREASURY}{a numeric vector}
\item{UNCOMMITTED.FUNDS}{a numeric vector}
\item{PAID.UP.CAPITAL}{a numeric vector}
\item{SHORT.TERM.DEBT}{a numeric vector}
\item{SALES}{a numeric vector}
\item{LABOR.COSTS}{a numeric vector}
\item{DEPRECIATION}{a numeric vector}
\item{OPERATING.PROFIT}{a numeric vector}
\item{FINANCIAL.OUTCOME}{a numeric vector}
\item{GROSS.PROFIT}{a numeric vector}
\item{NET.PROFIT}{a numeric vector}}}
\source{
Public use data from the CASC project.
}
\description{
A real data set comprising figures of 834 companies in the Tarragona area.
Data correspond to year 1995.
}
\examples{
data(Tarragona)
head(Tarragona)
dim(Tarragona)
}
\references{
Brand, R. and Domingo-Ferrer, J. and Mateo-Sanz, J.M., Reference
data sets to test and compare SDC methods for protection of numerical
microdata. Unpublished.
\url{http://neon.vb.cbs.nl/casc/CASCrefmicrodata.pdf}
}
\keyword{datasets}

