\name{testdata}
\alias{testdata}
\alias{testdata2}
\docType{data}
\title{
A real-world data set on household income and expenditures
}
\description{
A concise (1-5 lines) description of the dataset.
}
\usage{
data(testdata)
data(testdata2)
}
\format{
  A data frame with 4580 observations on the following 14 variables.
  \describe{
    \item{\code{urbrur}}{a numeric vector}
    \item{\code{roof}}{a numeric vector}
    \item{\code{walls}}{a numeric vector}
    \item{\code{water}}{a numeric vector}
    \item{\code{electcon}}{a numeric vector}
    \item{\code{relat}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{hhcivil}}{a numeric vector}
    \item{\code{expend}}{a numeric vector}
    \item{\code{income}}{a numeric vector}
    \item{\code{savings}}{a numeric vector}
    \item{\code{ori_hid}}{a numeric vector}
    \item{\code{sampling_weight}}{a numeric vector}
  }
  A data frame with 93 observations on the following 19 variables.
  \describe{
    \item{\code{urbrur}}{a numeric vector}
    \item{\code{roof}}{a numeric vector}
    \item{\code{walls}}{a numeric vector}
    \item{\code{water}}{a numeric vector}
    \item{\code{electcon}}{a numeric vector}
    \item{\code{relat}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{hhcivil}}{a numeric vector}
    \item{\code{expend}}{a numeric vector}
    \item{\code{income}}{a numeric vector}
    \item{\code{savings}}{a numeric vector}
    \item{\code{ori_hid}}{a numeric vector}
    \item{\code{sampling_weight}}{a numeric vector}
    \item{\code{represent}}{a numeric vector}
    \item{\code{category_count}}{a numeric vector}
    \item{\code{relat2}}{a numeric vector}
    \item{\code{water2}}{a numeric vector}
    \item{\code{water3}}{a numeric vector}
  }
}
\references{
The International Household Survey Network, www.ihsn.org
}
\examples{
data(testdata)
## maybe str(testdata) ; plot(testdata) ...
}
\keyword{datasets}